Basic tests for data
This adds in functions to process the data and turn it into python objects. From here, we'll turn it into a grid ready for Munkres to work out the answer. These tests are quite weak at the moment, because they give equal weight to all possible outcomes - so there are as many senior folks as there are junior. I'll need to work out some kind of weighting process for this