Remove matching module
I've pulled this module out into its own package. The aim of this is to make it easier to do different things with it - for example, put it in a web app like this or into a tKinter GUI app.

This is quite a significant change, but everything should still work