Implement Google OAuth flow
This brings in the authentication and authorisation flow through Google. The user is asked to login via Google and give access to their Drive.

TODO: Move the secrets from a local file into environment variables, or stored somewhere on the infrastructure platform, so that there's no risk of committing them by accident