Complete processing task
This commit completes the processing section of the service. It takes some files that have already been uploaded, processes them asynchronously with Celery, and then turn the result into the mailing list.

The tests should only be run in integration, as they require a running broker and backing service like Redis or RabbitMQ