Remove dialog code and placeholder for other code
I've removed `showDialog` and `closeDialog`.

I've replaced this with a `window.location.replace()` function to redirect to the new page.

@jonodrew I don't know how to handle adding the correct info the button itself.

I guess we'd need to pass something between the pages on the routing end, or we could use JS to look at the end of the URL and update the button on the download page. Thoughts?