Update cookie policy
As we're following best practice, this sets out my understanding the minimum requirement to meet the regulations under PECR & GDPR for cookies.

We only have to stipulate which cookies are used and whether they are essential or non-essential. If they are non-essential, we have to seek consent.

I understand we only use cookies for essential purposes, and so this should be sufficient.

@jonodrew can you fill in the blanks? What's the name you've given to the cookie that is set, and when does it expire?