Create mapping function to cope with broken package
The idiot who maintains the underlying package for this service has decided to generalise it, despite all the work we've done together. This means that our csv is no longer the template for the entire library, and I've had to write a mapping function to convert our headings into what they expect. Very frustrating but can't be helped.

Because I am the idiot.