Fixup breaking changes in 5.1.0
Users should not experience any degradation of the service. However, we may need to consider how best to move forwards with the creation of the spreadsheet and mailing list.

This change makes it our responsibility to turn the integers representing the grade to a human-readable string - which would be fine, if the library didn't also produce our spreadsheet, which needs to be human-readable. We'll need to take control of the spreadsheet, or prevail upon the maintainer to find some way for us to get a translating or mapping function into the processing.