Fix against upcoming deprecation
The underlying library is reverting its Person/Mentor/Mentee objects so that they only have a `profession` attribute. We're already passing "profession" as a keyword, but it's valuable for our Rule to be clear about what's being matched. So I've added a property on the CSMentor and the CSMentee classes