Introduce pickling
Pickle is Python's own serialisation library. It's not without its risks, but it does massively increase speed and enables passing around complex data objects - which is what we have here. It enables us to expose our "most mentees with a mentor" functionality.

To enable this I've added the `connections.setter` property in `CSPerson`, to allow pickle to load the data back into the object when it deserialises it. I've added some tests, but they need expanding - there's likely something around patching the long-running matching tasks to improve visibility of how everything is working.

The `run_task` route now takes a "pairing" variable, which indicates which function will be run to calculate the outcomes. The default is to use the quicker, one-round loop with an unmatched bonus of 6.