Lay out framework for notifications approach
Notifications are going to start with Notify, but may be expanded in the future. It might be better to start with something simpler, in that case...

There are tests to check that the async task is called the appropriate number of times, and that the task itself calls things the right way. We'll start by unit testing them, and then expand out into using the Notify sandbox to try out sending real emails